pyproject.toml
[project]
name = "bytebot-mvp"
version = "0.1.0"
description = "Karpathy-style mastery project for a Windows desktop automation agent."
authors = [{ name = "Bytebot Builder" }]
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
automation = [
  "pywinauto>=0.6.8",
  "uiautomation>=2.0.20",
  "pyautogui>=0.9.54",
  "pillow>=10.0.0",
]
dev = [
  "pytest>=8.3",
]

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["."]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.coverage.run]
branch = true
source = ["mvp"]
